requirements.txt
fastapi[standard]
uvicorn[standard]

# Dev tools
black
flake8
google-genai
